Update Produto.java
Alterado valores de float para double

+++ VendaMercadorias/src/entidade/Produto.java
@@ -11,16 +11,16 @@ package entidade;
 public class Produto {
     private int id;
     private String descricao;
-    private float valor;
-    private float quantidade;
+    private double valor;
+    private double quantidade;
 
-    public Produto(String descricao, float valor, float quantidade) {
+    public Produto(String descricao, double valor, double quantidade) {
         this.descricao = descricao;
         this.valor = valor;
         this.quantidade = quantidade;
     }
 
-    public Produto(int id, String descricao, float valor, float quantidade) {
+    public Produto(int id, String descricao, double valor, double quantidade) {
         this.id = id;
         this.descricao = descricao;
         this.valor = valor;
@@ -46,19 +46,19 @@ public class Produto {
         this.descricao = descricao;
     }
 
-    public float getValor() {
+    public double getValor() {
         return valor;
     }
 
-    public void setValor(float valor) {
+    public void setValor(double valor) {
         this.valor = valor;
     }
 
-    public float getQuantidade() {
+    public double getQuantidade() {
         return quantidade;
     }
 
-    public void setQuantidade(float quantidade) {
+    public void setQuantidade(double quantidade) {
         this.quantidade = quantidade;
     }
